move debug_one to base
