+++ river/facto/hofm.py
@@ -60,6 +60,13 @@ class HOFM(BaseFM):
         order_latents_dict = functools.partial(collections.defaultdict, random_latents)
         return collections.defaultdict(order_latents_dict)
 
+    def _interaction_names(self, x):
+        return [
+            " - ".join(map(str, combination))
+            for d in range(2, self.degree + 1)
+            for combination in itertools.combinations(x.keys(), d)
+        ]
+
     def _interaction_combination_keys(self, x):
         for d in range(2, self.degree + 1):
             for combination in itertools.combinations(x.keys(), d):
@@ -69,11 +76,10 @@ class HOFM(BaseFM):
         return functools.reduce(lambda x, y: x * y, (x[j] for j in combination))
 
     def _interaction_coefficient(self, combination):
-        d = len(combination)
         return sum(
             functools.reduce(
                 lambda x, y: np.multiply(x, y),
-                (self.latents[j][d] for j in combination),
+                (self.latents[j][len(combination)] for j in combination),
             )
         )
 
@@ -274,77 +280,6 @@ class HOFMRegressor(HOFM, base.Regressor):
     def predict_one(self, x):
         x = self._ohe_cat_features(x)
         return self._raw_dot(x)
-
-    def debug_one(self, x: dict, decimals: int = 5) -> str:
-        """Debugs the output of the fwfm regressor.
-        Parameters
-        ----------
-        x
-            A dictionary of features.
-        decimals
-            The number of decimals use for printing each numeric value.
-        Returns
-        -------
-        A table which explains the output.
-        """
-
-        x = self._ohe_cat_features(x)
-
-        def fmt_float(x):
-            return "{: ,.{prec}f}".format(x, prec=decimals)
-
-        names = (
-            [
-                " - ".join(map(str, combination))
-                for d in range(2, self.degree + 1)
-                for combination in itertools.combinations(x.keys(), d)
-            ]  # latents
-            + list(map(str, x.keys()))  # weights
-            + ["Intercept"]  # intercept
-        )
-
-        values = list(
-            map(
-                fmt_float,
-                [
-                    self._interaction_combination_vals(x, combination)
-                    for combination in self._interaction_combination_keys(x)
-                ]  # latents
-                + list(x.values())  # weights
-                + [1],  # intercept
-            )
-        )
-
-        weights = list(
-            map(
-                fmt_float,
-                [
-                    self._interaction_coefficient(combination)
-                    for combination in self._interaction_combination_keys(x)
-                ]  # latents
-                + [self.weights.get(i, 0) for i in x]  # weights
-                + [self.intercept],  # intercept
-            )
-        )
-        contributions = (
-            [
-                self._interaction_coefficient(combination)
-                * self._interaction_combination_vals(x, combination)
-                for combination in self._interaction_combination_keys(x)
-            ]  # latents
-            + [xi * self.weights.get(i, 0) for i, xi in x.items()]  # weights
-            + [self.intercept]  # intercept
-        )
-        order = reversed(np.argsort(contributions))
-        contributions = list(map(fmt_float, contributions))
-
-        table = utils.pretty.print_table(
-            headers=["Name", "Value", "Weight", "Contribution"],
-            columns=[names, values, weights, contributions],
-            order=order,
-        )
-
-        return table
 
 
 class HOFMClassifier(HOFM, base.Classifier):
